Reference

Reference

Representing map data

OpenStreetMapX.MapData — Type.

The MapData represents all data that have been processed from OpenStreetMap osm file This is the main data structure used fot map data analytics.

Fields

bounds : bounds of the area map (stored as a OpenStreetMapX.Bounds object)

nodes : dictionary of nodes representing all the objects on the map (with coordinates in East, North, Up system)

roadways : unique roads stored as a OpenStreetMapX.Way objects

intersections : roads intersections

g : LightGraphs directed graph representing a road network

v : vertices in the road network

e : edges in the graph represented as a tuple (source,destination)

w : edge weights, indexed by graph id

class : road class of each edge

source

OpenStreetMapX.get_map_data — Method.

get_map_data(filepath::String,filename::Union{String,Nothing}=nothing;
             road_levels::Set{Int} = Set(1:length(OpenStreetMapX.ROAD_CLASSES)),
			 use_cache::Bool = true, only_intersections::Bool=true)::MapData

High level function - parses .osm file and create the road network based on the map data.

Arguments

filepath : path with an .osm file (directory or path to a file)

filename : name of the file (when the first argument is a directory)

road_levels : a set with the road categories (see: OpenStreetMapX.ROAD_CLASSES for more informations)

use_cache : a *.cache file will be crated with a serialized map image in the datapath folder

only_intersections : include only road system data

source

OpenStreetMapX.LLA — Type.

LLA

Point in Latitude-Longitude-Altitude (LLA) coordinates Used to store node data in OpenStreetMapX XML files

Constructors

LLA(lat::Float64, lon::Float64)
LLA(lat::Float64, lon::Float64, alt::Float64)
LLA(xyz::XYZ)

Arguments

lat : lattitude

lon : Longitude

alt : altitude

source

OpenStreetMapX.ENU — Type.

ENU

Point in East-North-Up (ENU) coordinates.

Local cartesian coordinate system. Linearized about a reference point.

Constructors

ENU(east::Float64, north::Float64, up::Float64)
ENU(east::Float64, north::Float64)
ENU(xyz::XYZ)

source

OpenStreetMapX.Bounds — Type.

Bounds{T <: Union{LLA, ENU}}

Bounds for the LLA or ENUcoordinates.

source

Routing operations

generate_point_in_bounds(::MapData)
point_to_nodes(::Tuple{Float64,Float64}, ::MapData)
point_to_nodes(::LLA, ::MapData)
shortest_route(::MapData, ::Int, ::Int)
shortest_route(::MapData, ::Int, ::Int, ::Int)
fastest_route(::MapData, ::Int, ::Int, ::Dict{Int,Float64})
fastest_route(::MapData, ::Int, ::Int, ::Int, ::Dict{Int,Float64})

Google API routing

OpenStreetMapX.get_google_route — Method.

get_google_route(origin::Int, destination::Int,
                 map_data:MapData, googleapi_key::String;
                 googleapi_parameters::Dict{Symbol,String} = googleAPI_parameters)

Get route from to based on Google Distances API with two points (origin and destination) on map map_data using Google API key googleapi_key with optional Google Distances API request parameters googleapi_parameters.

source

OpenStreetMapX.get_google_route — Method.

get_google_route(origin::Int, destination::Int, waypoint::Int,
                 map_data:MapData, googleapi_key::String;
                 googleapi_parameters::Dict{Symbol,String} = googleAPI_parameters)

Get route from to based on Google Distances API with three points (origin, destination and waypoint between) on map map_data using Google API key googleapi_key with optional Google Distances API request parameters googleapi_parameters.

source

OpenStreetMapX.node_to_string — Method.

node_to_string(node_id::Int,map_data::MapData)

Convert node coordinates (stored in ENU system in the nodes field of map_data) identified by node_id to string with LLA system coordinates

source

OpenStreetMapX.googleAPI_parameters — Constant.

Dictionary for Google Distances API requests:

Keys

:url : url for google API, only JSON files outputs are accepted

:mode : transportation mode used in simulation, in the current library scope only driving is accepted

:avoid : map features to avoid (to mantain compatibility with OSM routes ferries should be avoided)

:units : unit system for displaing distances (changing to imperial needs deeper changes in both OSMsim and OpenStreetMapX modules)

source