Reference

Representing map data

OpenStreetMapX.MapData — Type

The MapData represents all data that have been processed from OpenStreetMap osm file This is the main data structure used fot map data analytics.

Fields

bounds : bounds of the area map (stored as a OpenStreetMapX.Bounds object)

nodes : dictionary of nodes representing all the objects on the map (with coordinates in East, North, Up system)

roadways : unique roads stored as a OpenStreetMapX.Way objects

intersections : roads intersections

g : LightGraphs directed graph representing a road network

v : vertices in the road network

e : edges in the graph represented as a tuple (source,destination)

w : edge weights, indexed by graph id

class : road class of each edge

OpenStreetMapX.get_map_data — Method

get_map_data(filepath::String,filename::Union{String,Nothing}=nothing;
             road_levels::Set{Int} = Set(1:length(OpenStreetMapX.ROAD_CLASSES)),
			 use_cache::Bool = true, only_intersections::Bool=true)::MapData

High level function - parses .osm file and create the road network based on the map data.

Arguments

filepath : path with an .osm file (directory or path to a file)

filename : name of the file (when the first argument is a directory)

road_levels : a set with the road categories (see: OpenStreetMapX.ROAD_CLASSES for more informations)

use_cache : a *.cache file will be crated with a serialized map image in the datapath folder

only_intersections : include only road system data

trim_to_connected_graph: trim orphan nodes in such way that the map is a strongly connected graph

Coordinate systems

OpenStreetMapX.ECEF — Type

ECEF

Point in Earth-Centered-Earth-Fixed (ECEF) coordinates. Global cartesian coordinate system rotating with the Earth.

Constructors

ECEF(x::Float64, y::Float64, z::Float64)

OpenStreetMapX.LLA — Type

LLA

Point in Latitude-Longitude-Altitude (LLA) coordinates Used to store node data in OpenStreetMapX XML files

Constructors

LLA(lat::Float64, lon::Float64)
LLA(lat::Float64, lon::Float64, alt::Float64)
LLA(xyz::XYZ)

Arguments

lat : lattitude

lon : Longitude

alt : altitude

OpenStreetMapX.ENU — Type

ENU

Point in East-North-Up (ENU) coordinates.

Local cartesian coordinate system. Linearized about a reference point.

Constructors

ENU(east::Float64, north::Float64, up::Float64)
ENU(east::Float64, north::Float64)
ENU(xyz::XYZ)

OpenStreetMapX.Bounds — Type

Bounds{T <: Union{LLA, ENU}}

Bounds for the LLA or ENUcoordinates. If T is not given Bounds{ENU} will be created.

OpenStreetMapX.center — Function

center(bounds::Bounds{ENU})

Get Center Point of ENU Bounds Region

center(bounds::Bounds{LLA})

Get Center Point of LLA Bounds Region

OpenStreetMapX.inbounds — Function

inbounds(loc::ENU, bounds::Bounds{ENU})

Check Whether a location loc is within bounds bounds

inbounds(loc::LLA, bounds::Bounds{LLA})

Check whether a location loc is within bounds bounds

OpenStreetMapX.onbounds — Function

onbounds(loc::T, bounds::Bounds{T}) where T<:Union{LLA,ENU}

Check whether a location loc is onbounds bounds Works only for points that have passed the inbounds test

OpenStreetMapX.latlon — Function

latlon(m::MapData,map_g_point_id::Int64)::Tuple{Float64, Float64}

Returns a tuple of lattitute and longitude for a given graph node identifier map_g_point_id in graph m.g (i.e. map_g_point_id ∈ 1:nv(m.g)).

Routing operations

OpenStreetMapX.generate_point_in_bounds — Method

generate_point_in_bounds(m::MapData)

Generates a random pair of Latitude-Longitude coordinates within boundaries of map m

OpenStreetMapX.point_to_nodes — Method

point_to_nodes(point::Tuple{Float64,Float64}, m::MapData)

Converts a pair Latitude-Longitude of coordinates point to a node on a map m The result is a node indentifier.

OpenStreetMapX.point_to_nodes — Method

point_to_nodes(point::LLA, m::MapData)

Converts a pair of coordinates LLA (Latitude-Longitude-Altitude) point to a node on a map m The result is a node indentifier.

OpenStreetMapX.shortest_route — Function

shortest_route(m::MapData, node1::Int, node2::Int; routing::Symbol = :astar)

Find Shortest route between node1 and node2 on map m.

shortest_route(m::MapData, node1::Int, node2::Int, node3::Int; routing::Symbol = :astar)

Find Shortest route between node1 and node2 and node3 on map m.

OpenStreetMapX.fastest_route — Function

fastest_route(m::MapData, node1::Int, node2::Int;
                    routing::Symbol = :astar,
                    speeds::Dict{Int,Float64}=SPEED_ROADS_URBAN)

Find fastest route between node1 and node2 on map m with assuming speeds for road classes.

fastest_route(m::MapData, node1::Int, node2::Int, node3::Int;
                    routing::Symbol = :astar,
                    speeds::Dict{Int,Float64}=SPEED_ROADS_URBAN)

Find fastest route between node1 and node2 and node3 on map m with assuming speeds for road classes.

OpenStreetMapX.a_star_algorithm — Function

a_star_algorithm(g::AbstractGraph{U},  
                s::Integer,                       
                t::Integer,                       
                distmx::AbstractMatrix{T}=LightGraphs.weights(g),
                heuristic::Function = (u,v) -> zero(T)) where {T, U}

High level function - implementation of A star search algorithm: (https://en.wikipedia.org/wiki/A*searchalgorithm). Based on the implementation in LightGraphs library, however significantly improved in terms of performance.

Arguments

g : graph object

S : start vertex

t : end vertex

distmx : distance matrix

heuristic : search heuristic function; by default returns zero

a_star_algorithm(m::MapData,  
                s::Integer,                       
                t::Integer)

A star search algorithm with straight line distance heuristic

Arguments

m : MapData object

S : start vertex

t : end vertex

distmx : distance matrix

OpenStreetMapX.distance — Function

distance(a::ENU, b::ENU)

Calculates a distance between two points a and b

distance(a::ECEF, b::ECEF)

Calculates a distance between two points a and b

Compute the distance of a route for some nodes data

Get Distances Between Edges

Google API routing

OpenStreetMapX.get_google_route — Method

get_google_route(origin::Int, destination::Int,
                 map_data:MapData, googleapi_key::String;
                 googleapi_parameters::Dict{Symbol,String} = googleAPI_parameters)

Get route from to based on Google Distances API with two points (origin and destination) on map map_data using Google API key googleapi_key with optional Google Distances API request parameters googleapi_parameters.

OpenStreetMapX.get_google_route — Method

get_google_route(origin::Int, destination::Int, waypoint::Int,
                 map_data:MapData, googleapi_key::String;
                 googleapi_parameters::Dict{Symbol,String} = googleAPI_parameters)

Get route from to based on Google Distances API with three points (origin, destination and waypoint between) on map map_data using Google API key googleapi_key with optional Google Distances API request parameters googleapi_parameters.

OpenStreetMapX.node_to_string — Method

node_to_string(node_id::Int,map_data::MapData)

Convert node coordinates (stored in ENU system in the nodes field of map_data) identified by node_id to string with LLA system coordinates

OpenStreetMapX.googleAPI_parameters — Constant

Dictionary for Google Distances API requests:

Keys

:url : url for google API, only JSON files outputs are accepted

:mode : transportation mode used in simulation, in the current library scope only driving is accepted

:avoid : map features to avoid (to mantain compatibility with OSM routes ferries should be avoided)

:units : unit system for displaing distances (changing to imperial needs deeper changes in both OSMsim and OpenStreetMapX modules)

OpenStreetMapX.encode_one — Function

Encode single coordinate (multiplied by 1e5 and rounded) in Google Polyline Algorithm Format

Arguments

val : single coordinate (multiplied by 1e5 and rounded)

OpenStreetMapX.encode — Function

Encode coordinates in Google Polyline Algorithm Format

Arguments

coords : coordinates in LLA system stored as a tuple

OpenStreetMapX.decode_one — Function

Decode single coordinate

Arguments

polyline : coordinates in Google Polyline Algorithm Format stored as an array of characters

index : position of each single coordinate in polyline array

OpenStreetMapX.decode — Function

Decode coordinates in Google Polyline Algorithm Format

Arguments

polyline : string containing coordinates in Polyline Algorithm Format

Routing parameters

OpenStreetMapX.ROAD_CLASSES — Constant

ROAD_CLASSES

Road classes. Information used for routing and plotting.

Ordered by typical significance

OpenStreetMapX.CYCLE_CLASSES — Constant

CYCLE_CLASSES

Cycle classes - Level 1: Bike paths - Level 2: Separated bike lanes (tracks) - Level 3: Bike lanes - Level 4: Bikes typically allowed but not specified

OpenStreetMapX.PED_CLASSES — Constant

PED_CLASSES

Pedestrain paths

- Level 1: Cycleways, walking paths, and pedestrian streets
- Level 2: Sidewalks
- Level 3: Pedestrians typically allowed but unspecified
- Level 4: Agricultural or horse paths, etc.

OpenStreetMapX.SPEED_ROADS_URBAN — Constant

Default Speed Limits in Kilometers Per Hour in urban areas

OpenStreetMapX.SPEED_ROADS_RURAL — Constant

Default Speed Limits in Kilometers Per Hour in rural areas